[FEAT] Added hit feedback

+++ Assets/Scripts/Entities/Player/Upgrades/UpgradeLevel.cs
@@ -7,6 +7,6 @@ public class UpgradeLevel
     // FIELDS
     [Min(0)] public int cost = 10;
     [Tooltip("Value at the CURRENT upgrade level")]
-    public int value = 1;
+    public float value = 1f;
     public string description;
 }
